Improve trigger platform typing
